requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
uvicorn[standard]==0.13.4
gunicorn==20.0.4
typing
fastapi
pandas
pydantic
python-multipart
google-cloud-bigquery